chore(supabase): updates database workspace architecture
