+++ bot/package.json
@@ -19,7 +19,7 @@
     "cron": "^2.2.0",
     "dayjs": "^1.11.7",
     "discord.js": "^14.7.1",
-    "supabase": "workspace:*"
+    "database": "workspace:*"
   },
   "devDependencies": {
     "@types/cron": "^2.0.0",
