+++ database/docs/04-permissions.generated.sql
