+++ database/types/main.types.ts
@@ -1,0 +1,6 @@
+import { SupabackTypes } from './generic.types';
+
+export type DraftyConfig = SupabackTypes<'drafty_configurations'>;
+export type EnrollmentMessage = SupabackTypes<'enrollment_messages'>;
+export type GameTable = SupabackTypes<'game_tables'>;
+export type Pod = SupabackTypes<'pods'>;
